% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.r
\docType{data}
\name{example_data}
\alias{example_data}
\title{Stan DDM Example Data}
\format{An object of class \code{data frame}.}
\usage{
StanDDM::example_data
}
\description{
A small example data set which has been simulated. First the parameters with with 
the function \code{\link{makeFakeParams}}, then, the data itself with the function
\code{\link{simulDat}}. It contains 10 simulated subjects with 10 'trials' each.
Use this data to test a new model or in general, the packages' functionalities.
}
\keyword{datasets}
