% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_data_processing.r
\name{experimental_data_processing}
\alias{experimental_data_processing}
\title{Experimental Data Processing}
\usage{
experimental_data_processing(a)
}
\arguments{
\item{a}{Data frame or tibble. Mind the required columns and names. 
See the example data set with \code{StanDDM::example_data}}
}
\value{
Returns a list with two entries: "forstan" is the data in the correct format to be used in fitting. 
"forsim" is for data simulation.
}
\description{
Read in and process experimental data for model fitting. It is assumed that there is at least 1 
response per side (left/right)! Else, this function will not work!
The column structure of the data has to be following:
SUBJECT_ID, REACTION TIME, STIMULI SIDE (alternatively ANSWER SIDE), CORRECTNESS OF RESPONSE
Mind that the data needs to be in a long and tidy format. The data processing is inspired by Ahn et al.'s 
procedure.
}
\references{
hBayesDM Package and paper: \url{https://www.ncbi.nlm.nih.gov/pubmed/29601060}
}
