% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{fake_data_processing}
\alias{fake_data_processing}
\title{Data Processing for Simulations}
\usage{
fake_data_processing(data)
}
\arguments{
\item{simulated_data}{Data frame or tibble.}
}
\value{
Returns a list with two entries: "forstan" is the data in the correct format to be used in fitting. "forsim" is for data simulation.
}
\description{
Processes output from data simulation functions for posterior use in model validation procedures. Only package internal use.
}
