% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{fit_quality}
\alias{fit_quality}
\title{Data Fit Assessment}
\usage{
fit_quality(experim_dat, simul_dat, model_name = "NAME_UNDEFINED")
}
\arguments{
\item{experim_dat}{Experimental data in the form produced by \code{\link{experimental_data_processing}}.}

\item{experim_dat}{Simulated data in the form produced by \code{\link{simulDat}}.}
}
\value{
Prints a table with RMSE and \eqn{R^2} values.
}
\description{
Calculates \eqn{R^{2}} values for correct and incorrect mean reaction times and RMSE values
for the proportions of correct answers per reaction times cuantiles.
}
