% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{init_packages}
\alias{init_packages}
\title{Initialize All Required Packages}
\usage{
init_packages()
}
\description{
A function initializing the needed functions. Only required when not installing the package.
}
