% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFakeParams.r
\name{makeFakeParams}
\alias{makeFakeParams}
\title{Pseudo-Random DDM Parameter Generation}
\usage{
makeFakeParams(nsub = 1, include = NULL)
}
\arguments{
\item{nsub}{Number of "simulated" subjects.}

\item{include}{A vector with the desired inter-trial variabilities.
'sv' for Drift Rate inter-trial variability.
'sz' for Bias inter-trial variability.
'st' for Non-Decision Time inter-trial variability.}
}
\value{
Returns a Data Frame with the desired parameters.
}
\description{
Generates a data_frame with pseudo-randomly generated DDM parameters for model
validation or similar.
}
\examples{

makeFakeParams(10, include = c('sv', 'st'))
}
