% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{model_diagnostics}
\alias{model_diagnostics}
\title{Stanfit MCMC Sampling Diagnostics}
\usage{
model_diagnostics(fit)
}
\arguments{
\item{experim_dat}{Experimental data in the form produced by \code{\link{experimental_data_processing}}.}

\item{experim_dat}{Simulated data in the form produced by \code{\link{simulDat}}.}

\item{model_name}{Optional name for the model to distinguish to which data/model the
function was applied to. Default is 'NAME_UNDEFINED'.}
}
\value{
Prints a table with RMSE and \eqn{R^2} values.
}
\description{
A wrapper function around several model diagnostics provided by \code{rstan}. It prints
the amount of divergent transitions, treedepth limit hits, energy,
}
