% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_judgement.r
\name{model_judgement}
\alias{model_judgement}
\title{Model Comparison via LOO-CV, WAIC and Raw LPPD}
\usage{
model_judgement(..., lik_name = "log_lik", impute_inf = TRUE)
}
\arguments{
\item{lik_name}{Name under which the log likelihoods have been saved in the models. Needs to be identical 
across all Stanfit objects.}

\item{impute_inf}{A boolean which regulates if underflow values should be automatically imputed or not. If
\code{FALSE}, the models with such values will just be ignored. If \code{TRUE}, a report will be generated
on how many values were imputed and for which models.}

\item{stanfits}{At least 2 stanfit objects.}
}
\value{
Prints a table and generates a plot with the model ranked according to LOO-CV, WAIC and Raw LPPD.
}
\description{
Applies log-likelihood based model comparison to any number of stanfit objects and extracts LOO-CV, 
WAIC and Raw LPPD measures and plots the models on a absolute scale ranging from 0 (best models) to 
1 (worst models). WAIC and LOO-CV code has been adapted directly from the "LOO" package. Because
most methods to prevent log-exp calculation-underflow failed, this function imputes underflow values
with the mean of sampled values. When this happens, the function will report the amount of imputed values
to inform the user. More than 5% of imputed values should not be accepted.
Mind how Stan saves log-likelihood values! Consult the Stan manual to check how they are saved correctly!
Because all models need to be loaded into memory, be wary if loading to big stanfits. Lighten the Stanfit
objects by discarding unnecesary iterations is advised.
}
\examples{

load(fit1.Rdata)
load(fit2.Rdata)

model_judgement(fit1, fit2, impute_inf = TRUE)
}
