% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_plots.r
\name{models_plots}
\alias{models_plots}
\title{Data Simulation Plots}
\usage{
models_plots(experim_dat, simul_dat, model_name = "NAME_UNDEFINED")
}
\arguments{
\item{experim_dat}{Experimental data in the form produced by \code{\link{experimental_data_processing}}.}

\item{model_name}{Optional name for the model to distinguish to which data/model the
function was applied to. Default is 'NAME_UNDEFINED'.}

\item{experim_dat}{Simulated data in the form produced by \code{\link{simulDat}}.}
}
\value{
Generates and saves plots in a directory with name of \code{model_name}.
}
\description{
Plots simulated and experimental data. The created plots are: Reaction times densities for 
correct and incorrect responses and the proportion of correct answeres for each one of 
six reaction time bins. A folder with the model name is created and the plots saved 
automatically in the current working directory.
}
