% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_recovery.r
\name{param_recovery}
\alias{param_recovery}
\title{Parameter Recovery for Model Testing without Experimental Data}
\usage{
param_recovery(fit, n_subjects, auth_params,
  model_name = "NAME_UNDEFINED")
}
\arguments{
\item{fit}{Stanfit object fitted on synthetic data.}

\item{n_subjects}{Number of "Simulated Subjects".}

\item{auth_params}{"Ground-truth" parameters used to simulate the data. In the form of the
output from the function \code{\link{makeFakeParams}}.}

\item{model_name}{Optional name for the model to distinguish the plots and created
folder easier. Default is 'NAME_UNDEFINED'.}
}
\value{
Creates a folder and saves plots which compare recovered parameters and "ground-truth" parameters
for each iteration of the MCMC sampler.
}
\description{
This function extracts DDM parameters from a Stanfit object run on simulated data. Then,
it and compares those parameters fitted on simulated data with the parameters used to simulate
the data in the first place. This function is intended to check if the model is capable to
"recover" parameters which are known. Plots for visual inspection are made automatically.
Automatic fit diagnostics yet to be implementd with this function. The comparison is done
for each parameter over all iterations to detect convergence problems or to spot difficult parameter
spaces.
}
