% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{parameter_extraction}
\alias{parameter_extraction}
\title{Automatic DDM Parameter Extraction}
\usage{
parameter_extraction(fit, numsub, names = NULL)
}
\arguments{
\item{names}{Subject names (Vector of strings) for correct plotting. Is not required, but recommended.}

\item{stanfit_object}{Output of a previous stanfit.}

\item{num_subjects}{Amount of fitted subjects. Required for automatic extraction.}
}
\value{
Returns a CSV file with all parameter estimates.
}
\description{
Automatically extracts parameter estimates from a stanfit object.
}
