% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precheck.r
\name{precheck}
\alias{precheck}
\title{Precheck for Underflow Values Imputation}
\usage{
precheck(arg, lik_name)
}
\arguments{
\item{arg}{A Stanfit object fitted on synthetic data.}

\item{lik_name}{Name under which the log likelihoods have been saved in the models. Needs to be identical}

\item{current_model}{Name of the current model (Currently not in use.)}
}
\value{
A list with information regarding the underflow values for each Stanfit object.
}
\description{
Checks Stanfit objects for underflow values before applying the \code{\link{model_judgement}} function.
}
