% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error (RMSE)}
\usage{
rmse(df)
}
\arguments{
\item{data_frame}{A data frame that contains simulated and experimental data, one in 
each column.}
}
\value{
Returns the RMSE value as double.
}
\description{
Calculates the RMSE between simulated and experimental data, stored in a data frame.
This is a helper function for \code{\link{fit_quality}}.
}
