% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif2.r
\name{runif2}
\alias{runif2}
\title{Random Value Generation: Uniform Distribution Scale-Location Reparametrization}
\usage{
runif2(n, loc, scale)
}
\arguments{
\item{n}{Amount of values to be generated.}

\item{loc}{Location parameter.}

\item{scale}{Scale parameter.}
}
\value{
A vector of doubles of length "n".
}
\description{
A long title for a small function. Helper function for \code{\link{drift_simuls}} which reparametrizes the arguments of
the \code{\link{runif}} function as scale and location parameters.
}
