% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveit.r
\name{saveit}
\alias{saveit}
\title{Save R Objects With External Name}
\usage{
saveit(..., string, file)
}
\arguments{
\item{string}{A string used as filename for the saved R object}

\item{file}{Any R object in the environment of any type or class.}
}
\description{
A wrapper function around \code{\link[base]{save}}. Assign a string as filename for a R object, 
instead the environment name. Will save file to disk.
}
