% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_template.r
\name{show_template}
\alias{show_template}
\title{Model Template}
\usage{
show_template()
}
\value{
Displays a model template in text form, to be copied and modified.
}
\description{
Displays a template for creating custom models. They can be added to the "include_models"
argument in the \code{\link{StanDDM}} function (one at the time). They need first to be loaded in
memory, then, the name has to be included in the argument.
}
