% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{simulDat}
\alias{simulDat}
\title{Data Generation From DDM Parameters}
\usage{
simulDat(...)
}
\arguments{
\item{parameter_data_frames}{A list of data frames that contain DDM parameters in
the format returned by the 'makeFakeParams' function: 1 column per parameter 
and as many rows as subjects. Can also be applied to just one data frame!
The rows can be named (in a data frame, in tibbles it is deprecated).}
}
\value{
Returns a list of Data Frames with simulated data. Or alternatively, a list of 1
data frame.
}
\description{
A wrapper around the function 'drift_simuls' which simulates as many data sets as 
data frames are provided in a list. This is a convenience function and 
the function 'drift_simuls' can also be used directly on just one set on parameters.
}
\examples{

parameters_1 #parameters extracted from a fitting procedure
parameters_2 #simulated parameters

sims <- simulDat(list(parameters_1, parameters_2))

sim_1 <- sims[[1]]
sim_2 <- sims[[2]]
}
