% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{StanDDM}
\alias{StanDDM}
\title{StanDDM: Multi-level Bayesian Fitting Procedure for Decision-Making Data}
\usage{
StanDDM(data = NULL, include_models = c(), num_cores = 4,
  simulation = FALSE, warmup = 500, iter = 2000, stepsize = 0.15,
  adapt_delta = 0.85, max_treedepth = 30)
}
\arguments{
\item{data}{If simulations are not done (\code{simulation = FALSE}), then a data frame with decision-making
data has to be provided. Required columns are 'suj' (subject ID), 'rt' (reaction times in seconds), 'crit' for the 
diffusion criteria (accuracy coding or stimuli coding, more info on the github readme) and 'cor' (correct or 
incorrect answer). See the github readme for a example.}

\item{include_models}{A vector of strings to indicate which models should be fitted to the data or used in 
the simulations. 'Pure', 'st', 'sv', 'sz', 'sv_sz', 'st_sv', 'sz_st' and 'sv_sz_st' can be included in any order.
It is also possible to load in custom models. See the github readme for instructions.}

\item{num_cores}{Number of CPU cores to be used for the fitting procedure. Note that for each core, a MCMC
chain will be assigned automatically. The minimum amount of available cores on your machine should be 4.}

\item{simulation}{A boolean that if \code{simulation = TRUE}, a parameter recovery study for the current model is launched.
In that mode, a set of parameters will be randomly generated and also the corresponding data for 10 simulated 
subjects. That synthetic data will then be fitted with the indicated model. Plots and fitting quality will be
saved in your working directory. Interesting are the by-iteration plots for each parameter.
If \code{simulation = FALSE}, a data frame has to be provided in the correct format. The models will fit
the provided data.}

\item{warmup}{Warmup iterations which will be discarded. 500 to 1000 are recommended, but should be adjusted
to accomodate rig capacities and model properties.}

\item{iter}{MCMC iterations per chain. 1000 to 5000 are recommended but should be adjusted
to accomodate rig capacities and model properties.}

\item{stepsize}{Initial NUTS step-size. Will adapt after a some iterations. Adjust to own parameter space
if necessary. See Stan manual for more information.}

\item{adapt_delta}{See Stan manual for more information.}

\item{max_treedepth}{See Stan manual for more information.}
}
\value{
Returns a list of Data Frames with simulated data. Or alternatively, a list of 1
}
\description{
This is the main function, which calles all models defined in "include" and applies all Stan sampler
arguments to the fitting.
}
