% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.r
\name{waic}
\alias{waic}
\title{LOO-CV, WAIC and Raw LPPD Calculations}
\usage{
waic(stanfit, current_model, lik_name, impute_inf)
}
\arguments{
\item{stanfit}{A Stanfit object fitted on synthetic data.}

\item{current_model}{Name of the current model (Currently not in use.)}

\item{lik_name}{Name under which the log likelihoods have been saved in the models. Needs to be identical 
across all Stanfit objects.}

\item{impute_inf}{A boolean which regulates if underflow values should be automatically imputed or not. If
\code{FALSE}, the models with such values will just be ignored. If \code{TRUE}, a report will be generated
on how many values were imputed and for which models.}
}
\value{
A list with LOO-CV, WAIC and Raw LPPD calculations.
}
\description{
Implements LOO-CV, WAIC and Raw LPPD for the \code{\link{model_judgement}} function. Contains
the helper function "colVars" which calculates row-wise variances efficiently.
}
