% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental_data_processing.r
\name{experimental_data_processing}
\alias{experimental_data_processing}
\title{Experimental Data Processing}
\usage{
experimental_data_processing(a)
}
\arguments{
\item{experimental_data}{Data frame or tibble. Mind the required columns and names.}
}
\value{
Returns a list with two entries: "forstan" is the data in the correct format to be used in fitting. "forsim" is for data simulation.
}
\description{
Read in and process experimental data for model fitting. It is assumed that there is at least 1 
response per side (left/right)! Else, this function will not work and throw a 
"replacement has length zero" error.
The column structure of the data has to be following:
SUBJECT_ID, REACTION TIME, STIMULI SIDE (alternatively ANSWER SIDE), CORRECTNESS OF RESPONSE
Mind that the data needs to be in a long and tidy format.
As of now, this function has to be manually modified to change to stimuli fitting.
}
