% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.r
\name{model_diagnostics}
\alias{model_diagnostics}
\title{Stanfit MCMC Sampling Diagnostics}
\usage{
model_diagnostics(fit)
}
\arguments{
\item{fit}{Any stanfit object.}
}
\value{
Prints model diagnostics to console.
}
\description{
A wrapper function around several model diagnostics provided by \code{rstan}. It prints
the amount of divergent transitions, treedepth limit hits, energy, effective transitions 
and R-hat.
}
